+++ lib/karafka/pro/contracts/consumer_group_topic.rb
@@ -22,10 +22,30 @@ module Karafka
           ).fetch('en').fetch('validations').fetch('pro_consumer_group_topic')
         end
 
-        virtual do |data|
+        nested(:virtual_partitions) do
+          required(:active) { |val| [true, false].include?(val) }
+          required(:partitioner) { |val| val.nil? || val.respond_to?(:call) }
+          required(:concurrency) { |val| val.is_a?(Integer) && val >= 1 }
+        end
+
+        virtual do |data, errors|
+          next unless errors.empty?
           next if data[:consumer] < Karafka::Pro::BaseConsumer
 
           [[%i[consumer], :consumer_format]]
+        end
+
+        # When virtual partitions are defined, partitioner needs to respond to `#call` and it
+        # cannot be nil
+        virtual do |data, errors|
+          next unless errors.empty?
+
+          virtual_partitions = data[:virtual_partitions]
+
+          next unless virtual_partitions[:active]
+          next if virtual_partitions[:partitioner].respond_to?(:call)
+
+          [[%i[virtual_partitions partitioner], :respond_to_call]]
         end
       end
     end