+++ spec/integrations/pro/consumption/virtual_partitions/many_batches_same_key.rb
@@ -20,7 +20,9 @@ draw_routes do
   consumer_group DT.consumer_group do
     topic DT.topic do
       consumer Consumer
-      virtual_partitioner ->(_msg) { '1' }
+      virtual_partitions(
+        partitioner: ->(_msg) { '1' }
+      )
     end
   end
 end