+++ spec/integrations/pro/shutdown/virtual_partitions/shutdown_on_all_consumers.rb
@@ -25,7 +25,9 @@ draw_routes do
   consumer_group DT.consumer_group do
     topic DT.topic do
       consumer Consumer
-      virtual_partitioner ->(msg) { msg.raw_payload }
+      virtual_partitions(
+        partitioner: ->(msg) { msg.raw_payload }
+      )
     end
   end
 end